+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Changed
 - Change Utah Header wrapper from a `<div>` to a `<header>` element.
-- Have Icon Button show a ToolTip 
+- Have Icon Button show a ToolTip
+- Show title on icon button
 
 ## Deprecated